refactor(codec): split lib api support modules
